Move global_vars to utils
